this-and-arguments-support
Previously, references to `this` and `arguments` didn't work properly in jwacs,
because if they occur within a continuation function then they refer to the
`this` or `arguments` associated with the continuation call instead of with the
original function.  Eg, this code:

  function foo(one)
  {
    var x = getConfig('foo');
    var y = getConfig('bar');
    alert(arguments[0]);
  }

  foo('first arg');

will output 'first arg' if executed as straight Javascript.  However, the cps
transformation changes it into something more like:

  function foo($k, one)
  {
    getConfig(function JW0(x) {
      getConfig(function JW1(y) {
        alert(arguments[0]);
      }, 'bar')
    }, 'foo');
  }

  foo($id, 'first arg');

If we execute this transformed code with no further modifications, then it will
output whatever the second call to `getConfig` returns.  (ie, the first argument
to the JW1 continuation function, which is where the `alert` call happens to
appear after the transformation).

There's an equivalent problem with `this` references.

We solve this problem by saving a reference to `arguments` and/or `this` at the
beginning of the original function and changing all references to refer to this
"shadow" value instead.  We do this in the new SHADOW-VALUES transformation,
which runs just before the cps transformation.

The shadow-values transformation would translate the original source above into
the following:

  function foo(one)
  {
    var arguments$0 = arguments;
    var x = getConfig('foo');
    var y = getConfig('bar');
    alert(arguments$[0]);
  }

  foo('first arg');

which the cps transformation will translate to:

  function foo($k, one)
  {
    var arguments$0 = arguments;
    getConfig(function JW0(x) {
      getConfig(function JW1(y) {
        alert(arguments$0[0]);
      }, 'bar')
    }, 'foo');
  }

  foo($id, 'first arg');

This level of shadowing is good enough for `this` references, but unfortunately
there is an additional complication for `arguments` references: The extra
continuation argument.  This latest code will still not behave as expected,
because although it is now referring to the correct `arguments` object, it is
not referring to the correct argument, since `arguments[0]` is now `$k` instead
of `one`.

To resolve /that/ problem, I've also updated the RUNTIME transformation to
replace references to `arguments` with a call to a new runtime function called
`$makeArguments`, which generates a new arguments-style object that doesn't
include the continuation argument.  So the RUNTIME transformation will replace
`var arguments$0 = arguments;` with `var arguments$0 = $makeArguments(arguments);`.

There are a couple of other subtleties: We don't want to touch references to
user-declared variables that are named `arguments`, but it's the early
SHADOW-VALUES that determines whether such a declaration exists.  So it replaces
the identifier element for "real" `arguments` references with a special-value
element, and the RUNTIME transformation only adds `$makeArguments` to
special-value `arguments` references.

darcs-hash:20060608215551-53808-4a419bc7f8aeb03ebbca84e1d026f8115b757913