fix-cps-fn-recursion
The cps transformation wasn't recursing into the FN slots of FN-CALLs nor the
CONSTRUCTOR slots of NEW-EXPRs in the mistaken belief that the explicitize
transformation made that unnecessary.  However, it is possible for either of
those slots to contain a FUNCTION-EXPRESSION after explicitization, so recursion
is in fact necessary.

This patch adds the recursion and also modifies a test in lang-tests.jw to
expose the problem if it exists.

darcs-hash:20060609163557-53808-2d5fb8487a47bcf95b6e4816964297634208b984