fix-cps-new-expr-tail-returns
We were treating tail-returns of new expressions as simple value returns instead
of as a tail call to a function.  This fix adds the proper handling for new
expressions to the return method of the cps transformation.

Eg, this code:

  return new foo(10);

should become

  return new foo($k, 10);

and /not/:

  return $k(new foo(10));

darcs-hash:20060610192224-53808-c22b35d82ff0bcca3776f7dff3cf9b1024229932