fix-global-variable-visibility
Under normal circumstances, we transform a variable declaration whose
initializer is a function call into a function call with a continuation whose
argument has the variable's name, so that

  var x = foo();

becomes

  foo(function(x) { ... });

That works okay in local functions so long as nobody refers to a variable before
it is declared [1].  However, at the toplevel, /most/ global variables will be
referred to before they have been defined, because they will be referenced from
within functions, which have been shifted ahead of everything else to avoid
their being obscured by being embedded into continuation functions.  That means
that global variable declarations will be invisible to functions, which is
pretty obviously the Wrong Thing to do.  eg:

  var myGlobal = new String("hello");
  function foo()
  {
    alert(myGlobal);
  }

was being transformed into something akin to:

  function foo($k)
  {
    alert(function() { return $k(); }, myGlobal);
  }

  new String(function (myGlobal) { }, "hello");

In the transformed version above, `myGlobal` is no longer visible from `foo`.

In fact, this problem is very much akin to the problem that the
shift-function-decls transformation was invented to solve (that of function
declarations being obscured by getting buried in a continuation function).

So we solve it by generalizing shift-function-decls into shift-decls.  We add
the new behaviour that empty variable declarations are stuck at the top of the
toplevel scope for all global variables, and the locations of their former
initializations are replaced with simple assignment statements.

So, the new shift-decls transformation would transform the above code like so:

  var myGlobal;
  function foo()
  {
    alert(myGlobal);
  }

  myGlobal = new String("hello");

which will down the line be turned into this:

  var myGlobal;
  function foo($k)
  {
    alert(function() { return $k(); }, myGlobal);
  }

  new String(function (JW0) { myGlobal = JW0; }, "hello");

Notice that `myGlobal` is now set properly yet still accessible to `foo`.

-----
[1]
By requiring variables to be defined before they are used, we are breaking
Javascript semantics.  We should actually be applying this transformation to all
scopes, not just to the toplevel.  The problem is that that could wind up
doubling the number of local variables floating around (ie, the original
variable, plus the continuation parameter that it gets initialized from), which
seems likely to have horrible performance implications.

We need to fix this long-term, but I want to do it in an optimized way (where we
only shift variables that actually are referred to before they are declared), so
I'm going to do that as a separate check-in.

darcs-hash:20060615000925-53808-78975f5a41c65a1983906fc174f517813ed13aac