backchannel-messages
The thought of adding another multiple-value return to the TX-CPS function which
would only be looked at by a grand total of three methods but which would need
to be propagated by all methods filled me with dread.  As a procrastination
technique, I have implemented some constructs for signalling messages up the
call-chain using a "backchannel".

The basic idea is that we want to throw return values "past" intervening
function calls to the calling function that actually cares about it.  We do this
by signalling on named backchannels.  You signal a message using BACKCHANNEL-SIGNAL
and receive them using BIND-WITH-BACKCHANNELS. Eg:

  JW> (defun foo (n)
        (if (oddp n)
          (backchannel-signal :odd n)
          (backchannel-signal :even n)))
  FOO
  JW> (defun bar ()
        (bind-with-backchannels (final :odd odds)
            (dotimes (n 50 n)
              (foo n))
          (format t "~&bar: odds = ~S" odds)))

  BAR
  JW> (defun baz ()
        (bind-with-backchannels (final :odd odds :even evens)
            (bar)
          (format t "~&baz odds = ~S~%evens = ~S~%" odds evens)))
  BAZ
  JW> (bar)
  bar: odds = (1 3 5 7 9 11 13 15 17 19 21 23 25 27 29 31 33 35 37 39 41 43 45 47 49)
  NIL
  JW> (baz)
  bar: odds = (1 3 5 7 9 11 13 15 17 19 21 23 25 27 29 31 33 35 37 39 41 43 45 47 49)
  baz odds = NIL
  evens = (0 2 4 6 8 10 12 14 16 18 20 22 24 26 28 30 32 34 36 38 40 42 44 46 48)
  NIL

The list of signals received on a given channel is bound during the execution of
the body of BIND-WITH-BACKCHANNELS.  So for example, in BAR, a list of signals
received on the :ODD channel by BAR is bound to the variable ODDS.  The first
actual return value of the DOTIMES form is boudn to FINAL.  If we had listed
multiple return values, they would have been bound as well.  Eg, the form

  (BIND-WITH-BACKCHANNELS (V1 V2 :BC BC-MSGS)
      (QUUX)
    (BOO)
    (BOOP))

binds V1 and V2 to the first and second return values of QUUX (and binds BC-MSGS
to the list of messages received on the :BC channel while QUUX was executing).
We distinguish backchannel bindings (of the form :CHANNEL-NAME CHANNEL-BINDING)
from regular return value bindings (of the form BINDING) based on the fact that
channel names must be keywords and binding names must not be.  So everything
before to the first keyword is considered a regular return value binding;
everything afterward is considered a backchannel binding.

Note that backchannel messages do not propagate beyond the first handler that
collects the given channel.  So in the first example above, BAR receives the
:ODD channel messages from FOO, so BAZ does not, but since BAR does not handle
messages on the :EVEN channel, they propagate all the way to BAZ.

Signalling a backchannel message does not affect the flow of control in the
signalling function.  Either the message is received, in which case
BACKCHANNEL-SIGNAL returns T, or it is not, in which case BACKCHANNEL-SIGNAL
returns NIL.  Signalling a message that is not received does no harm.

darcs-hash:20060618203016-53808-99ab8884ec19b48e6d529ff938e3df26c9cd3b25