fix-cps-labelled-continuations
Labelled continuations can cause problems because they reorder code, which means
that a var decl that gets turned into a continuation parameter may no longer be
visible to code in a labelled continuation that references it.

We fix this problem by detecting these cases and changing the var decl to an
assignment to a variable that has a "bare" declaration at the beginning of the
enclosing scope (either function or toplevel).

darcs-hash:20060618210610-53808-64cbb4343bff5a01574e31abe2072d2c17ff6695