fix-toplevel-effective-fn-calls
Function calls and new expressions (aka "effective function calls") were broken
at the toplevel when the continuation contained one or more other effective
function calls, because we assume when doing the trampoline transformation of
these effective function calls in the continuation that a pogo loop (viz. the
`$trampoline` runtime function) is somewhere in the call stack.

This assumption is true within functions, because all functions have the "call
guard" that invokes `$trampoline` (through `$callByDirect`) if they have not
been called in the expected way.  However, this assumption was false for the
first call at the toplevel.

We now wrap the first toplevel call in a call to `$trampoline` so that the boxed
values that are returned by calls within the continuation are treated properly.

darcs-hash:20060619224036-53808-2d7dfd30420df7eec61ad5214ef8feba18ae11cd