default-transformation-flattening
The default transform methods now perform "flattening".  What this means is that
for any given element that expects one of its slots to be a single statement, the
default processing will now automatically wrap the results of a transformation of
that slot in a statement-block if a list of statements comes back.

What _that_ means is that we will now do the right thing if any "obligate
statement" element is converted from a single statement into a list of
statements by a transformation.  The CPS transformation already did this in a
kind of half-assed fashion; now all transformations can without too much extra
effort.

darcs-hash:20060622184237-53808-872d9bf9e269e150410272b5c15e25979c58e1f8