fix-toplevel-trampoline-interface
We were getting a weird problem where running lang-tests worked, but the console
always said "ret has no properties".  The problem was that, since the toplevel
is not a function, it was not guaranteed to be explicitly terminated (and in
fact will most often not be).  That means that the last function to run in the
toplevel will return `undefined`, which causes the trampoline loop to crap out.

The reason that the tests still worked was that the toplevel code runs while the
page is loading, whereas the main function runs _after_ the page is finished
loading (by being called from the onLoad event handler).  So the toplevel would
crap out at the very end (after it had already done all its work) and drop a
complaint into the console.  Then the onLoaded event fires and the main function
starts grinding away.

We fix this problem by explicitly terminating the toplevel with a `suspend`
statement (much as we explicitly terminate function bodies with a `return`
statement).  This is only necessary if the toplevel is both unterminated and
also contains at least one effective function call (since effective function
calls are the only reason we would ever run a trampoline at the toplevel).

This patch introduces the new file static-analysis.lisp, which contains some
control-flow functions that did not otherwise have an obvious home.  Currently
it contains EXPLICITLY-TERMINATED-P (previously in type-analysis.lisp) and the
new INTRODUCES-FUNCTION-CALL-P.

darcs-hash:20060628032851-53808-65b92a57c61dfad70c60d14caa7e5a2fd6dd8384