fix-shift-var-decls
The shift-decls transformation was correctly creating empty variable
declarations for all the var-decls in the toplevel scope, but it wasn't
stripping decls that occurred inside of other statements (eg, in a try body), so
we were winding up with duplicated variable declarations.  Which isn't
syntactically incorrect, but it's pretty weak to be generating that kind of code.

I've added the STRIP-VAR-DECLS-IN-SCOPE helper transformation to delve into
substatements and find all the variable declarations that need to be stripped
(ie, turned into assignment statements).

darcs-hash:20060628181537-53808-24be8d9e8d05cf8d50ee54406d5322cb3900db11