switch-to-thunk-based-handler-stacks/source-model-and-cps-transform
This is phase 1 of the switch from using a global variable to manage the
exception handler stacks to using a local variable in the $trampoline function.

I refer to this new scheme as "thunk-based", because instead of calling a
runtime function that adjusts a global variable in order to add/remove exception
handlers, clients will now return special boxed values that indicate to the
$trampoline function that such adjustments need to be made.  These boxed values
will contain a field that indicates the handler stack adjustment to make, and a
thunk field that tells the trampoline how to resume the computation.

To support this change, ADD-HANDLER and REMOVE-HANDLER source elements now need
to be created with a thunk-body slot that contains the transformed
statement-tail immediately after the add/remove operation.  This patch makes that
change.

A subsequent patch will include the changes to the runtime functions and
runtime/trampoline transformations.

darcs-hash:20060629001725-53808-51c4d5c74646c70fe58ff8fb17a444a3636ea382