switch-to-thunk-based-handler-stacks/runtime-and-trampoline-transforms
This is phase 2 (the final phase) of the switch from using a global variable to
manage the exception handler stacks to using a local variable in the $trampoline
function.

In the new thunk-based scheme, we communicate changes in the handler stack by
returning special boxed thunks instead of by calling runtime functions, so most
of the runtime transform handling for ADD-HANDLER and REMOVE-HANDLER source
elements has been removed.  The runtime transform now does no special handling
for `resume` and `suspend` statements, either.

The trampoline transformation is now in charge of converting ADD-HANDLER and
REMOVE-HANDLER source elements into boxed thunks with addHandler and
removeHandler properties.  We no longer need the REPLACE-HANDLER-STACK
administrative source element, since `resume` statements are converted into
special boxed thunks by the trampoline transformation, without needing to
communicate the need for a runtime function call to the runtime transform.

The Javascript runtime has changed substantially, of course:

  - The `$handlerStack` variable and the `$addHandler`, $removeHandler`, and
    `$replaceHandlers` functions have been removed
  - The `$trampoline` function now checks all boxed values for the presence of
    addHandler, removeHandler, and replaceHandler properties, and manages its
    local handlerStack variable accordingly.
  - The handlerStack variable is passed to all thunks as an argument, to allow
    continuations that are captured within the thunk's scope to point to the
    correct handler stack.  This parameter has a standard name (`$e`), which is
    also the name of a variable created at the beginning of the function to point
    to the $exHandlers property of the incoming continuation.  (This is guaranteed
    to be the correct stack state at the beginning of the function, of course, since
    we haven't had time to add any handlers of our own yet).
  - The `$makeK` runtime function now accepts a handler stack as its second argument
    rather than grabbing the current value of the local variable.  (We always pass
    `$e` to this function, which will pass it either the most recent enclosing thunk's
    parameter or the variable at the beginning of the function).

For the sake of consistency, we no longer allow the option of passing an array
of handlers to remove to the REMOVE-HANDLER element; when multiple nested
handlers are being removed, a series of REMOVE-HANDLER elements are now chained
together.

darcs-hash:20060704210129-53808-8af4c149f88bec278e51cd52eb8e0a95847f264b