extend-runtime
Extending the runtime to use IFRAME-based history management:

  - We now emit a default blank.html instead of a default bootstrap.html
    during compilation
  - The runtime function `$call0` now reports errors in a friendlier fashion
    (TODO: extend to `$call`).
  - Added IFRAME-based history management functions to JwacsLib

This is just a stopgap checkin.  The final history-management scheme will be
slightly different; I'm going to swallow my discomfort an implement an RSH-like
scheme that depends on a timer event that watches the value of the current
location.

darcs-hash:20060711053056-53808-71d3de1338b50958850d41773ee41624d2a34215