fix-strip-var-decls-of-simple-initializations
The strip-var-decls handling in the cps transformation had a bug where a
variable that was

  a) an escaping reference, and
  b) initialized with a simple expression (eg, `day < 1` or `20`)

would be improperly stripped.  eg, the following statement (assuming that `evar`
is an escaping reference)

  var evar = day < 1;

was being converted to

  var evar;
  ...
  evar = JWn;

(where `JWn` is not defined anywhere) instead of to

  var evar;
  ...
  evar = day < 1;

The problem was a simple error where the assignment-statement that is pre-
generated for effective-function-call initializations was used instead of a
separate assignment statement that includes the simple initialization
expression.

We now generate a proper assignment statement for that case.

This patch also includes a regression unit test and lang-test.

darcs-hash:20060714195128-53808-86d87296dc335fbcfba5c0f22123ae0efb7357d6