throw-into-continuations
The `resume` statement allows the programmer to resume a continuation, making it
appear as if the function that the continuation was captured from has just
returned.

We now add syntax to the `throw` statement to allow the programmer to throw an
exception into a continuation, making it appear as if the function that the
continuation was captured from has thrown.  The statement

  throw foo -> bar;

will throw the exception `foo` into the continuation `bar`.  The standard
arrowless syntax `throw foo;` behaves as usual (ie, throws `foo` into the
current continuation).

eg:

  var k = null;
  function grabK()
  {
    k = function_continuation;
    suspend;
  }

  try
  {
    grabK();
    alert('hello');
  }
  catch(e)
  {
    alert('exception ' + e);
    suspend;
  }

  try
  {
    throw 42 ->

darcs-hash:20060714224511-53808-32c858bdf7c86daf09574745aae1b7840fc5a407