move-template-imports-to-end-of-body
The original template had the <% jwacs_imports %> line in the <head> section.
However, that means that you can't see the "Loading..." line until all of the
scripts have loaded, and really that's the part that takes the time anyway.  So
now we put the script tags at the end of the <body> section, so that the user at
least has a little bit of feedback while all the hundreds of kilobytes of text
loads.

darcs-hash:20060718212358-53808-91cb1c063b13b937d7633eb0e0336319f5824aaa