fix-strip-var-decls-for-new-initializer
Due to a typo, we were using the wrong name for the continuation argument in the
new-expr initializer case for var-decl statements that need to be stripped.
(Whew!)  This is a similar sort of typo as I fixed in the patch
fix-strip-var-decls-of-simple-initializations.

This patch includes regression tests in both the unit test suite and also lang-tests.

darcs-hash:20060720012633-53808-0f3204e5d42e24f5c8589ec8f49e5f858817d6fa