add-yacc-restarts
Added INSERT-TERMINAL, REPLACE-TERMINAL, and SKIP-TERMINAL restarts to the yacc
parser.  This allows the user to manually fiddle with the lexer stream a bit
when an error occurs, which (funny coincidence) turns out to be exactly what we
need to be able to do in order to perform semicolon-insertion processing.

darcs-hash:20060825154007-53808-0fbf9358e8fe5b3b8fc025c9b10ab63cda77a2cd