implement-semicolon-insertion-1
The parser now implements semicolon insertion as defined by the spec.  We trap
errors coming out of the parser, and invoke the INSERT-TERMINAL restart to
insert semicolons when the conditions laid out by the spec are met.

The lexer now provides another function which will return T if the latest lexeme
was preceded by a LineTerminator.  It's getting to the point where I think the
lexer should actually be a CLOS object which we define methods against; the
"return several closures as multiple return values" routine is getting
complicated as the number of functions that we need grows.

I've only implemented this for the yacc parser for the time being, so I've
switched the *use-yacc* build parameter to always be T.  Hopefully I'll be able
to implement something similar for parsergen soon so that I can go back to using
the Lispworks parser, because the YACC parser is currently taking upwards of a
minute to compile on my machine.

I've added a new condition type (SYNTAX-ERROR) which is signalled whenever we
can't fix things using semicolon insertion.

NOTE: This is /not/ a complete implementation of the standard.  The standard has
cases where semicolons should be inserted before non-offending tokens (whenever
there's a newline in a "no newline" production); that is doable but annoying and
will take some additional lexer magic, so I'm deferring it.  There are some unit
tests flagged as expected failures for dealing with these cases.

darcs-hash:20060825165028-53808-69fe2a940c321464110ca0deb2a02dc602fa581a