implement-semicolon-insertion-2
We now have a complete implementation of semicolon insertion.  The "restricted
productions" are handled by creating a class of "restricted tokens" for which
the lexer will emit additional :LINE-TERMINATOR/:NO-LINE-TERMINATOR tokens.

The :BREAK, :CONTINUE, :RETURN, and :THROW tokens are :POST-restricted, which
means the lexer will emit :LINE-TERMINATOR after them if they are followed by a
line terminator, or :NO-LINE-TERMINATOR if they are not.

The :PLUS2 and :MINUS2 tokens are :PRE-restricted, which means the lexer will
emit :LINE-TERMINATOR after them if they are /preceded/ by a line terminator, or
:NO-LINE-TERMINATOR if they are not.

So, (lex "break;") ==> (:BREAK :NO-LINE-TERMINATOR :SEMICOLON),
    (lex "continue
          foo") ==> (:CONTINUE :LINE-TERMINATOR :IDENTIFIER), and
    (lex "a ++ b") ==> (:IDENTIFIER :NO-LINE-TERMINATOR :PLUS2 :IDENTIFIER).

Once we have the extra line-termination tokens, modifying the grammar to include
them in the restricted productions is straightforward.

darcs-hash:20060825195657-53808-4371864a6ad97f64939e76e11643d37b8c64b28c