fix-unget
The unget queue should actually be an unget stack, because if I call the
INSERT-TERMINAL restart 3 times with arguments A, B, and C, then C should be
inserted /before/ B and A (that's what "insert" means!).

So, this checkin converts the unget queue to an unget stack.

darcs-hash:20060825214949-53808-580b5f221521f56d24ba1c13137fc857acf99ea0