fix-re-literal-parsing
The lexer consumes regular expression literals in preference to single slashes
whenever possible.  However, this is occasionally NOT what we want.  Eg:

  var y = x/10 + x/20;

Based on the observation that most parser contexts can accept a slash or an
re-literal but not both, I've added another case to the parser's error-recovery
scheme.  If the parser chokes on an re-literal, reset the lexer to the start of
the re and consume a slash instead.

This deals with 3 of the 4 possible cases:
  1) re is acceptable but not slash: lexes the re in the first place
  2) slash is acceptable but not re: chokes on re, recovers by lexing slash instead
  3) neither is acceptable: lexes re, chokes, recovers to slash, chokes, signals error
  4) both is acceptable: UH OH

in case 4 we will prefer an RE, whereas the spec claims we should prefer a
slash.  Oh well.  I can't actually think of a situation where that would make
sense, so I can't write a test case, so whatver, we're still better off than we
were.  In practice the problems have all been in the area of case 2, which we
now handle just fine.

Getting this working required adding the ability for the lexer to back up to a
previous position.  Having MAKE-JAVASCRIPT-LEXER return yet another lambda in
yet another multiple return value was just getting disgusting, so I've factored
the lexer to be a bunch of methods on a CLOS object as part of this change.

darcs-hash:20060825224457-53808-18d9d2d663ec5f354442d09cd161487bebe76ab8