implement-inline-output-mode
We now provide an inline-output mode for BUILD-APP.  The user provides an output
stream, and the entire app is written to a single HTML file which is output to
the stream.

Files on disk may still be compiled if they are imported.

darcs-hash:20060828213959-53808-ed919c5d48eb125d7b7bbcfc29f3db1c520283a4