improve-$call0-error-reporting
$call0 now adds additional error information to exceptions only if the exception
is thrown in the actual function, not if it's thrown during the continuation
call.

This change only applies to calls to direct-style functions; however, that
accounts for the vast majority of calls to $call0, so it's worthwhile.  (Most
calls to cps-functions are inlined, because it's easier for us to tell that a
cps function is in cps).

darcs-hash:20060906222136-53808-89ecee74a08f854f1552b5a59778313bff246802