fix-invoke-handler
The "Invoke Handler" link was broken; it was causing a "ret has no properties"
exception.

The problem was that we were calling $trampoline with an argument
`function() { resume handler <- e; }`.  The problem with that is that the thunk
gets transformed into a thunk that accepts a continuation and starts up a new
`$trampoline` instance if it doesn't get one, which plays havoc with our
assumptions.

Fixed by adding an `$invokeHandler` function to the debug runtime that calls
`$trampoline` with the non-transformed thunk that it is expecting.

darcs-hash:20060908225650-53808-073898c53ccbcd79dedfba3c9102d690a0913aee