fix-loop-handling-of-nested-var-decls
The loop transformation was producing lists of var-decls from single var decls
in loop bodiest, which is a big problem when the var decl in question is the
initializer of a for loop.  Now we produce single statements where possible
(without resorting to statement blocks).

darcs-hash:20060920040003-53808-29c8a3b22ad1ec66ab68da32d2b56fbcd3659088