Harness for correctness testing in npm [WIP]
This implements a test runner capable of running transformed and
original versions of a function side-by side and diffing the effect
traces.

The idea is, to be able to quickly verify behavioral equivalence of a
CPS - transformed program, to its original, over varying (possibly
generated) sets of inputs.

the interface for a test is

function(CONT, F, G, H, ..., M) {
  function(a, b, c, ..., h) {
    // << code >>
  }
}

where F, G, H, ..., M are effects available to the test code
and a, b, c, ..., h are test case input values

Effect Interface:

Effects are regular functions, that would generally record their
calling order, inputs and outputs to an effect trace and provide a
programmable return value.

In testing of transformed programs, effects may also program their
continuation(s) and suspend.

TBD

Swapping out number-gen-async proved to be inconsistent. The traces
have identical heads, until the (equivalently seeded) prng
diverges. Whether this is due to an (undesirable) reordering of
traces, or due to some problem with the prng isolation in the test
suite is to be determined ..

Larger (possibly generated) suits of program fragments, excercising
JS' features could be created ...